pyproject.toml
[project]
name = "osint-deep-research"
version = "0.1.0"
description = "A deep research framework for OSINT with hierarchical validation and LLM-as-Judge voting"
authors = [
    {name = "Nathan Daeila", email = "nathanbdaeila@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0.0"
dependencies = ["langgraph (>=1.0.5,<2.0.0)", "langchain-core (>=1.2.3,<2.0.0)", "langchain-openai (>=1.1.6,<2.0.0)", "python-dotenv (>=1.2.1,<2.0.0)", "langfuse[openai] (>=3.11.2,<4.0.0)", "langchain (>=1.2.0,<2.0.0)", "numpy (>=2.4.0,<3.0.0)", "selenium (>=4.39.0,<5.0.0)", "browser-use (>=0.11.2,<0.12.0)", "tavily-python (>=0.7.17,<0.8.0)", "uv (>=0.9.18,<0.10.0)", "google-search-results (>=2.4.2,<3.0.0)", "ruff (>=0.14.10,<0.15.0)", "langchain-google-genai (>=4.1.2,<5.0.0)"]

[project.scripts]
odr = "odr.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "odr"}]

[tool.poetry.group.dev.dependencies]
# Dev dependencies will be added via `poetry add --group dev`
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
  "integration: marks tests as live/integration (deselect with '-m \"not integration\"')",
]
